First working implementation of RooSTLRefCountList.
