+++ roofit/roofitcore/inc/LinkDef.h
@@ -3,4 +3,4 @@
 #include "roofit/roofitcore/inc/LinkDef3.h"
 #include "roofit/roofitcore/inc/LinkDef4.h"
 
-#pragma link C++ class RefCountListNew<RooAbsArg>+;
+#pragma link C++ class RooSTLRefCountList<RooAbsArg>+;
